+++ DBReport/My Project/Settings.Designer.vb
@@ -533,6 +533,18 @@ Namespace My
                 Me("DisplayFieldLength") = value
             End Set
         End Property
+        
+        <Global.System.Configuration.UserScopedSettingAttribute(),  _
+         Global.System.Diagnostics.DebuggerNonUserCodeAttribute(),  _
+         Global.System.Configuration.DefaultSettingValueAttribute("False")>  _
+        Public Property DisplayBigIntAndSmallIntAsInteger() As Boolean
+            Get
+                Return CType(Me("DisplayBigIntAndSmallIntAsInteger"),Boolean)
+            End Get
+            Set
+                Me("DisplayBigIntAndSmallIntAsInteger") = value
+            End Set
+        End Property
     End Class
 End Namespace